pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawless"
version = "0.2.0"
description = "Safety-first agent framework — everything is a skill"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
voice = [
    "vosk>=0.3",
    "piper-tts>=1.2",
    "sounddevice>=0.4",
    "numpy>=1.24",
]
faiss = [
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.2",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
cl-bot = "clawless.user.main:main"
cl-admin = "clawless.admin.main:main"

[tool.setuptools]
packages = [
    "clawless",
    "clawless.user",
    "clawless.user.skills",
    "clawless.user.skills.cli",
    "clawless.user.skills.reasoning",
    "clawless.user.skills.memory",
    "clawless.user.skills.proposer",
    "clawless.admin",
]

[tool.setuptools.package-dir]
clawless = "src"

[tool.ruff]
line-length = 100
target-version = "py311"
